requirements.txt
dotenv
fastapi
google-genai
supabase
streamlit
uvicorn